[boot-test] Include a default QEMU boot shim in product bundles
For general QEMU support it is convenient to include the default QEMU
boot shim when one is not explicitly specified (unless we expect to boot
from an EFI disk image). This is already the behaviour of the main
product bundle.

Bug: 304279485
Change-Id: Ibe5e72afdce15315363be177af26b2bf8d1b7392
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/979200
Fuchsia-Auto-Submit: Joshua Seaton <joshuaseaton@google.com>
Reviewed-by: Ina Huh <ihuh@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>