[netstack3] Move inspect_devices to DeviceIpApi
When getting ready to move `inspect_devices`, I realized that the amount
of information we were exposing from the IP layer was really lacking.
The next realization was that to expose more information with our
current inspect pattern I'd have to either create many duplicate types
to expose to bindings or expose the internal types, and in both cases
add a lot of noise to the types exported to bindings.

Instead, introduce a new pattern where we abstract the Fuchsia inspector
instead and perform all the inspecting in core. This should massively
decrease the cost of adding new inspect data, which has been a burden
and was discouraging adding more data to inspect.

Bug: b/42083910
Change-Id: I9458e416d9a5334eeb0fa173a61f45e1f2988ba7
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/979737
Fuchsia-Auto-Submit: Bruno Dal Bo <brunodalbo@google.com>
Reviewed-by: Jeff Martin <martinjeffrey@google.com>
Reviewed-by: Amanda Tait <atait@google.com>
Commit-Queue: Bruno Dal Bo <brunodalbo@google.com>