[gvisor_syscall_tests] Use test sharding framework
Uses //src/lib/testing/sharding to shard syscall tests to avoid
the whole suite timing out.

As an initial pass, every suite is sharded into 5 parts, where
shard_part_regex = "([^/]+(?:/[^/]+)?).*" in order to shard based on
the first two slash-delimited segments of the name (or the whole name
if there are no slashes).

As a result, we can remove the elevated timeout_secs and allow the TCP
suite to run on riscv.

Fixed: 318529127
Change-Id: I9de493afa7afea6415858c6b1ab1ba8472304b00
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/979234
Commit-Queue: Evan Wildenhain <wildenhain@google.com>
Reviewed-by: Bruno Dal Bo <brunodalbo@google.com>