[assembly] Remove package size check merge
This feature is no longer being used, and we have no plans to use it.
This was originally used to create a budget for the base package (system
image) which was a merge of many other packages, but we removed this
budget, because nearly all of the packages were removed from the input,
and the base package is no longer something interesting to check.

Bug: b/321066699
Change-Id: I8ea491568ce597b41fa5e0a08363044c304a7f9f
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/980422
Reviewed-by: David Iserovich <diserovich@google.com>
Fuchsia-Auto-Submit: Aidan Wolter <awolter@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>